requirements.txt
# streamlit
requests-cache
pandas
openmeteo-requests
retry-requests
folium
streamlit-folium
geopy
geocoder
tensorflow
sklearn
xgboost
